requirements.txt
jupyter
keras_tqdm
matplotlib
pandas
pytest
subword-nmt
tensorflow-gpu
tqdm
